requirements.txt
streamlit==11.37.1
tensorflow==2.17.0
Pillow==10.3.0
numpy==1.26.4
